backend/requirements.txt
azure-cognitiveservices-vision-face==0.5.0
azure-common==1.1.27
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
importlib-metadata==4.8.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.1
Pillow==8.3.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.5.1
